+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   * "setLocation(String countryCode, String city, String location, String ipAddress)"
   * "setDisableLocation()"
 * Crash Reporting interface added and accessible through "Countly::instance()::crash()" call.
+* Added "disableUnhandledCrashReporting" function to the "Config" class to disable automatic uncaught crash reporting.
+* Added "setMaxBreadcrumbCount(int)" function to the "Config" class to change allowed max breadcrumb count.
 
 * Fixed a bug where setting custom user properties would not work.
 * Fixed a bug where setting organization of the user would not work.
@@ -17,7 +19,6 @@
 * Deprecated "Countly::backendMode()" call, use "Countly::backendM" instead via "instance()" call.
 * Deprecated "Usage::addLocation(double, double)" call, use "Countly::location::setLocation" instead via "instance()" call.
 * Deprecated "Usage::addCrashReport()" call, use "Countly::crash" instead via "instance()" call.
-* Deprecated "Countly::backendMode()" call, use "Countly::backendM" instead via "instance()" call.
 * The following methods are deprecated from the "UserEditor" interface:
   * "commit()" instead use "Countly::userProfile::save" via "instance()" call
   * "pushUnique(String, Object)" instead use "Countly::userProfile::pushUnique" via "instance()" call